[Backend] Refactor leaderboard logic into adhoc
