+++ backend/src/adhoc/boards/__init__.py
@@ -1,0 +1,2 @@
+from .base import Board
+from .config import get_boards
