+++ backend/src/state/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from .announcement_state import Announcement, Announcements
-from .board_state import Board, FirstBloodBoard, ScoreBoard, SimpleScoreBoard
 from .game_policy_state import GamePolicy
 from .game_state import Game
 from .hint_state import Hint, Hints
